\name{AsExample}
\alias{AsExample}
\alias{AsExample}
\docType{data}
\title{Example arsenic concentrations in drinking water}
\usage{data(AsExample)}
\source{None. Generated.}
\description{
Artificial numbers representing arsenic concentrations in a drinking
water supply.

Objective is to determine what can be done with data where all values
are below the reporting limit.  There is a detection limit at 1, and a
reporting limit at 3 ug/L.  Used in Chapter 8 of the NADA book
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


