\name{asSurv-methods}
\docType{methods}
\alias{asSurv}
\alias{asSurv-methods}
\alias{asSurv,Cen-method}
\alias{asSurv,formula-method}
\title{Methods for function asSurv in Package NADA}
\description{
    Methods for function \code{asSurv} in package \pkg{NADA}.

    \code{asSurv} converts a \code{Cen} object to a \code{Surv} object.
}
\usage{
\S4method{asSurv}{Cen}(x)

\S4method{asSurv}{formula}(x)
} % End usage

\arguments{
\item{x}{A \code{Cen} or \code{formula} object.}
} % End arguments
\keyword{methods}

