\name{Atra}
\alias{Atra}
\docType{data}
\title{Atrazine concentrations in Nebraska ground water}
\usage{data(Atra)}
\source{
Junk et al., 1980, Journal of Environmental Quality 9, pp.
479-483.}
\description{
Atrazine concentrations in a series of Nebraska wells before (June)
and after (September) the growing season.

Objective is to determine if concentrations increase from June to September.
There is one detection limit, at 0.01 ug/L.  Used in Chapters 4, 5,
and 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


