\name{AtraAlt}
\alias{AtraAlt}
\alias{AtraAlt}
\docType{data}
\title{Atrazine concentrations in Nebraska ground water -- Alternative}
\usage{data(AtraAlt)}
\source{
Altered from the data of Junk et al., 1980, Journal of Environmental
Quality 9, pp. 479-483.}
\description{
Alternative Atrazine concentrations altered from the Atra data set so
that there are more nondetects, adding a second detection limit at 0.05.

Objective is to determine if concentrations increase from June to September.
There are two detection limits, at 0.01 and 0.05 ug/L.  Used in Chapters
5 and 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}
