\name{Atrazine}
\alias{Atrazine}
\docType{data}
\title{Atrazine concentrations in Nebraska ground water -- Another Format}
\usage{data(Atrazine)}
\source{
Junk et al., 1980, Journal of Environmental Quality 9, pp.
479-483.}
\description{
The same atrazine concentrations as in Atra, stacked into one column
(col.1).  Column 2 indicates the month of collection.  Column 3 indicates
which data are below the detection limit those with a value of 1.

Objective is to determine if concentrations increase from June to September.
There is one detection limit, at 0.01 ug/L.  Used in Chapter 9 of the
NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


