\name{Bloodlead}
\alias{Bloodlead}
\docType{data}
\title{Lead concentrations in the blood of herons in Virginia.}
\usage{data(Bloodlead)}
\source{
Golden et al., 2003, Environmental Toxicology and Chemistry 22,
1517-1524.}
\description{
Blood-lead concentrations in herons of Virginia.  Objective is to compute
interval estimates for lead concentrations.  There is one detection limit,
at 0.02 ug/g.  Used in Chapter 7 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


