\name{boxplot-methods}
\docType{methods}
\alias{boxplot-methods}
\alias{boxplot,ros-method}
\alias{boxplot,ros}
\title{Methods for function boxplot in Package NADA}
\description{
  Methods for box plotting objects in package \pkg{NADA} 
}
\usage{
    \S4method{boxplot}{ros}(x, ...)
} % usage

\arguments{

\item{x}{
    An output object from a NADA function such as \code{\link{ros}}.
}
\item{...}{
        Additional arguments passed to the generic \code{boxplot} method.
}
} % arguments
\seealso{
    \code{\link{boxplot}}
}
\keyword{methods}
