\name{Cadmium}
\alias{Cadmium}
\docType{data}
\title{Cadmium concentrations in fish}
\usage{data(Cadmium)}
\source{
none.  Data modeled after several reports.}
\description{
Cadmium concentrations in fish for two regions of the Rocky Mountains.

Objective is to determine if concentrations are the same or different
in fish livers of the two regions.  There are four detection limits,
at 0.2, 0.3, 0.4, and 0.6 ug/L.  Used in Chapter 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


