\name{cendiff-methods}
\docType{methods}
\alias{cendiff-methods}
\alias{cendiff,formula,missing,missing-method}
\alias{cendiff,Cen,missing,missing-method}
\alias{cendiff,numeric,logical,factor-method}
\title{Methods for function cendiff in Package NADA}
\description{
%    Methods for function \code{cendiff} in package \pkg{NADA}.
    See \code{\link{cendiff}} for all the details.
}
%\usage{
%
%\S4method{cendiff}{numeric,logical,factor}(obs, censored, groups, ...)
%
%\S4method{cendiff}{formula}(obs, ...)
%
%} % End usage
%
%\arguments{
%
%\item{obs}{
%    Either a numeric vector of observations or a formula.
%    See examples in \code{\link{cendiff}}.
%} 
%\item{censored}{
%    A logical vector indicating TRUE where an observation in `obs' is
%    censored (a less-than value) and FALSE otherwise.
%}
%\item{groups}{
%    A factor vector used for grouping `obs' into subsets.
%}
%\item{...}{
%    Addtional items that are common to this function and the \code{survfit}
%    function from the `survival' package.  See Details.
%}
%} % End arguments
%
%\seealso{
%    \code{\link{cendiff}}, \code{\link{Cen}}, \code{\link{survdiff}}
%}
%
\keyword{methods}
