\name{cenken-methods}
\docType{methods}
\alias{cenken-methods}
\alias{cenken,numeric,logical,numeric,logical-method}
\alias{cenken,numeric,missing,numeric,logical-method}
\alias{cenken,numeric,numeric,logical,missing-method}
\alias{cenken,numeric,Cen,missing,missing-method}
\alias{cenken,Cen,Cen,missing,missing-method}
\title{Methods for function cenken in Package NADA}
\description{
    See \code{\link{cenken}} for all the details.
}
\keyword{methods}
