\name{cenmle-methods}
\docType{methods}
\alias{cenmle-methods}
\alias{cenmle,Cen,missing,missing-method}
\alias{cenmle,numeric,logical,missing-method}
\alias{cenmle,numeric,logical,factor-method}
\alias{cenmle,numeric,logical,numeric-method}
\title{Methods for function cenmle in Package NADA}
\description{
    See \code{\link{cenmle}} for all the details.
}
\keyword{methods}
