\name{cenreg-methods}
\docType{methods}
\alias{cenreg-methods}
\alias{cenreg,numeric,logical,numeric-method}
\alias{cenreg,numeric,logical,factor-method}
\alias{cenreg,formula,missing,missing-method}
\title{Methods for function cenreg in Package NADA}
\description{
    See \code{\link{cenreg}} for all the details.
}
\keyword{methods}
