\name{censummary-methods}
\docType{methods}
\alias{censummary-methods}
\alias{censummary,numeric,logical,missing-method}
\alias{censummary,numeric,logical,factor-method}
\alias{censummary,numeric,logical,numeric-method}

%\alias{censummary,Cen,missing,missing-method}
%\alias{censummary,formula,missing,missing-method}

\title{Methods for function censummary in Package NADA}
\description{
    See \code{\link{censummary}} for all the details.
}
\keyword{methods}
