\name{ChlfmCA}
\alias{ChlfmCA}
\docType{data}
\title{Chloroform concentrations in California groundwater.}
\usage{data(ChlfmCA)}
\source{
Squillace et al., 1999, Environmental Science and Technology 33,
4176-4187.}
\description{
Chloroform concentrations in groundwaters of California.

Objective is to determine if concentrations differ between urban and rural
areas.  There are three detection limits, at 0.05, 0.1, and 0.2 ug/L.
Used in Chapter 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


