\name{cor-methods}
\docType{methods}
\alias{cor-methods}
\alias{cor,ANY-method}
\alias{cor,cenmle-method}
\title{Methods for function cor in Package NADA}
\description{
  Methods for function \code{cor} in package \pkg{NADA} 
}
\section{Methods}{
\describe{
\item{x = "cenmle"}{Extracts the r-likelihood correlation coefficient from
a \code{cenmle} object.}
}
} % descibe
\keyword{methods}
