\name{CuZn}
\alias{CuZn}
\docType{data}
\title{Copper and zinc concentrations in ground water}
\usage{data(CuZn)}
\source{
Millard and Deverel, 1988, Water Resources Research 24, pp.
2087-2098.}
\description{
Copper and zinc concentrations in ground waters from two zones in the
San Joaquin Valley of California.  The zinc concentrations were used.

Objective is to determine if zinc concentrations differ between the
two zones.  Zinc has two detection limits, at 3 and 10 ug/L.  Used in
Chapters 4, 5 and 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


