\name{CuZnAlt}
\alias{CuZnAlt}
\docType{data}
\title{Zinc concentrations of the CuZn data set}
\usage{data(CuZnAlt)}
\source{
Altered from the data of Millard and Deverel, 1988, Water Resources
Research 24, pp. 2087-2098.}
\description{
Zinc concentrations of the CuZn data set;  concentrations in the
Alluvial Fan zone have been altered so that there are more nondetects.
This produces a greater signal, even with more nondetects.

Objective is to determine if zinc concentrations differ between the
two zones.  Zinc has two detection limits, at 3 and 10 ug/L.  Used in
Chapter 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


