\name{DFe}
\alias{DFe}
\docType{data}
\title{Dissolved iron concentrations from the Brazos River, USA}
\usage{data(DFe)}
\source{
Hughes and Millard, 1988, Water Resources Bulletin 24, pp.
521-531.}
\description{
Dissolved iron concentrations over several years in the Brazos River,
Texas.	Summer concentrations were used.

Objective is to determine if there is a trend over time.  Iron has two
detection limits, at 3 and 10 ug/L.  Used in Chapters 5, 11 and 12 of
the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


