\name{DOC}
\alias{DOC}
\docType{data}
\title{DOC in ground water}
\usage{data(DOC)}
\source{
Junk et al., 1980, Journal of Environmental Quality 9, pp.
479-483.}
\description{
Dissolved Organic Carbon (DOC) concentrations in ground waters of
irrigated and non-irrigated areas.

Objective is to determine if concentrations differ between irrigated
and non-irrigated areas.  There is one detection limit at 0.2 ug/L.
Used in Chapter 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


