\name{Golden}
\alias{Golden}
\docType{data}
\title{Blood lead in organs of herons from Virginia}
\usage{data(Golden)}
\source{
Golden et al., 2003, Environmental Toxicology and Chemistry 22, pp.
1517-1524.}
\description{
Lead concentrations in the blood and several organs of herons in Virginia.

Objective is to determine the relationships between lead concentrations in
the blood and various organs.  Do concentrations reflect environmental
lead concentrations, as represented by dosing groups?  There is one
detection limit, at 0.02 ug/g.  Used in Chapters 10 and 11 of the
NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


