\name{Hatchery}
\alias{Hatchery}
\docType{data}
\title{Antibiotic concentrations in fish-hatchery drainage}
\usage{data(Hatchery)}
\source{
Thurman et al., 2002, Occurrence of antibiotics in water from fish
hatcheries. USGS Fact Sheet FS 120-02.}
\description{
Proportions of detectable concentrations of antibiotics (ug/L) in drainage
from fish hatcheries across the United States.

Objective is to compute confidence intervals and tests on proportions.

There is one detection limit for each compound, all at 0.05 ug/L.
Used in Chapters 8 and 9 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


