\name{HgFish}
\alias{HgFish}
\docType{data}
\title{Mercury concentrations in fish across the United States.}
\usage{data(HgFish)}
\source{
Brumbaugh et al., 2001, USGS Biological Science Report
BSR-2001-0009.}
\description{ 
Mercury concentrations in fish across the United States.

Objective is to determine if mercury concentrations differ by
watershed land use.  Can concentrations be related to water and sediment
characteristics of the streams?  

There are three detection limits, at 0.03, 0.05, and 0.10 ug/g wet weight.
Used in Chapters 10, 11 and 12 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


