\name{MDCu}
\alias{MDCu}
\docType{data}
\title{Copper in ground water from of San Joaquin Valley, USA}
\usage{data(MDCu)}
\source{
Millard and Deverel, 1988, Water Resources Research 24, pp.
2087-2098.}
\description{
Copper concentrations in ground water from the Alluvial Fan zone in
the San Joaquin Valley of California.  One observation was altered to
become a <21, larger than all of the detected observations (the largest
detected observation is a 20).

Objective is to calculate summary statistics when the largest observation
is censored.

There are five detection limits, at 1, 2, 5, 10 and 20 ug/L.
An additional artificial detection limit of 21 was added to illustrate
a point.  Used in Chapter 6 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


