\name{Oahu}
\alias{Oahu}
\docType{data}
\title{Arsenic concentrations in Manoa Stream, Oahu Hawaii}
\usage{data(Oahu)}
\source{
Tomlinson, 2003, Effects of Ground-Water/Surface-Water Interactions
and Land Use on Water Quality.	Written communication (draft USGS
report).}
\description{
Arsenic concentrations (ug/L) in an urban stream, Manoa Stream at Kanewai
Field, on Oahu, Hawaii.

Objective is to characterize conditions by computing summary statistics.

There are three detection limits, at 0.9, 1, and 2 ug/L.  Uncensored
values reported below the lowest detection limit indicate that informative
censoring may have been used, and so the results are likely biased high.
Used in Chapter 6 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


