\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,ANY-method}
\alias{print,Cen-method}
\alias{print,cenfit-method}
\alias{print,cenmle-method}
\alias{print,cenreg-method}
\alias{print,censummary-method}
\alias{print,summary.cenreg-method}
\alias{print,ros-method}
\alias{print,NADAList-method}
\title{Methods for function print in Package NADA}
\description{
  Methods for function \code{print} in package \pkg{NADA} 
}
\section{Methods}{
\describe{

\item{x = "ANY"}{Default print method}

\item{x = "Cen"}{Displays a \code{\link{Cen}} object.}

\item{x = "cenfit"}{Displays a \code{\link{cenfit}} object.}

\item{x = "cenmle"}{Displays a \code{\link{cenmle}} object.}

\item{x = "cenreg"}{Displays a \code{\link{cenreg}} object.}

\item{x = "summary.cenreg"}{Displays a \code{\link{summary.cenreg}} object.}

\item{x = "ros"}{Displays a \code{\link{ros}} object.}

\item{x = "censummary"}{Displays a \code{\link{censummary}} object.}

\item{x = "NADAList"}{Displays a \code{NADAList} object.  This is an
internal method and should rarely be used from the command line.}
}
} % descibe
\keyword{methods}
