\name{Recon}
\alias{Recon}
\docType{data}
\title{Atrazine in streams of the Midwestern U.S.}
\usage{data(Recon)}
\source{
Mueller et al., 1997, Journal of Environmental Quality 26, pp.
1223-1230.}
\description{
Atrazine concentrations in streams throughout the Midwestern United
States.

Objective is to develop a regression of model for atrazine concentrations
using explanatory variables.

There is one detection limit, at 0.05 ug/L.  Used in Chapter 12 of the
NADA book.  
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


