\name{Roach}
\alias{Roach}
\docType{data}
\title{Lindane in fish from tributaries of the Thames River, UK}
\usage{data(Roach)}
\source{
Yamaguchi et al., 2003, Chemosphere 50, 265-273. }
\description{
Lindane concentrations in fish from tributaries of the Thames River,
England.

Objective is to determine whether lindane concentrations are the same
at all sites.

There is one detection limit at 0.08 ug/kg.  Used in Chapter 9 of the
NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


