\name{ros-methods}
\docType{methods}
\alias{ros-methods}
\alias{ros,numeric,logical-method}
\title{Methods for function ros in Package NADA}
\description{
  Methods for constructing ROS models in package \pkg{NADA} 
}
\section{Methods}{
\describe{
\item{obs = "numeric", censored = "logical"}{
    Compute and return a ROS model
    given a numeric vector of observations and a logical vector 
    indicating TRUE or FALSE where the observations are not censored or
    censored respectively.
}
}}
\keyword{methods}
