\name{SedPb}
\alias{SedPb}
\docType{data}
\title{Lead in stream sediments before and after wildfires}
\usage{data(SedPb)}
\source{
Eppinger et al., 2003, USGS Open-File Report 03-152. }
\description{
Lead concentrations in stream sediments before and after wildfires.

Objective is to determine whether lead concentrations are the same pre-
and post-fire.

There is one detection limit at 4 ug/L.  Used in Chapter 9 of the
NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


