\name{ShePyrene}
\alias{ShePyrene}
\docType{data}
\title{Pyrene concentrations in water from Puget Sound, WA USA}
\usage{data(ShePyrene)}
\source{
She, N., 1997, Analyzing censored water quality data using a nonparametric
approach.  Journal of the American Water Resources Association, 33, pp615--624.
}
\description{
Pyrene concentrations in milligrams per liter from 20 water-quality
monitoring stations in the Puget Sound of Washington State, USA.

Used for characterizing priority pollutant concentrations in sediments
of Puget Sound by computing summary statisitics.  Contains eight detection
limits with 11 nondetects out of 56 total measurements.
}
\references{
She, N., 1997, Analyzing censored water quality data using a nonparametric
approach.  Journal of the American Water Resources Association, 33, pp615--624.

Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


