\name{Silver}
\alias{Ag}
\docType{data}
\title{Silver-standard concentrations}
\usage{data(Silver)}
\source{
Helsel and Cohn, 1988, Water Resources Research 24, pp. 1997-2004.}
\description{
Silver concentrations in a standard solution sent to 56 laboratories as
part of a quality assurance program.

Objective is to estimate summary statistics for the standard solution.
The median or mean might be considered the most likel estimate
of the concentration.

Contains twelve detection limits, the largest at 25 ug/L.  Used in
Chapter 6 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


