\name{Tbl1one}
\alias{Tbl1one}
\docType{data}
\title{Contaminant concentrations in test and a control group}
\usage{data(Tbl1one)}
\source{None. Generated data.}
\description{
Contaminant concentrations in test and a control group.

Objective is to determine whether a test group has higher concentrations
than a control group.

There are three detection limits, at 1, 2, and 5 ug/L.  Used in Chapter 1,
Table 1.1 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


