\name{TCE}
\alias{TCE}
\docType{data}
\title{TCE in ground waters of Long Island, New York}
\usage{data(TCE)}
\source{
Eckhardt et al., 1989,	USGS Water Resources Investigations Report
86-4142.
}
\description{
TCE concentrations (ug/L) in ground waters of Long Island, New York.
Categorized by the dominant land use type (low, medium, or high density
residential) surrounding the wells.

Objective determine if concentrations are the same for the three land
use types.  There are Four detection limits, at 1,2,4 and 5 ug/L.
Used in Chapter 10 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


