\name{TCEReg}
\alias{TCEReg}
\docType{data}
\title{TCE ground waters of Long Island -- with explanatory variables}
\usage{data(TCEReg)}
\source{
Eckhardt et al., 1989,	USGS Water Resources Investigations Report
86-4142.}
\description{
TCE concentrations (ug/L) in ground waters of Long Island, New York,
along with several possible explanatory variables.

Objective is to determine if concentrations are related to one or more
explanatory variables.

There are four detection limits, at 1,2,4 and 5 ug/L.  One column
indicates whether concentrations are above or below 5.  Used in Chapter
12 of the NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


