\name{Thames}
\alias{Thames}
\docType{data}
\title{Dieldrin, lindane and PCB in fish of the Thames River, UK}
\usage{data(Thames)}
\source{
Yamaguchi et al., 2003, Chemosphere 50, 265-273.}
\description{
Dieldrin, lindane and PCB concentrations in fish of the Thames River
and tributaries, England.

Objective is to determine if concentrations differ among sampling
sites.  Are dieldrin and lindane concentrations correlated?  There is
one detection limit per compound.  Used in Chapters 11 and 12 of the
NADA book.
}
\references{
Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}
