\name{cenken-methods}
\docType{methods}
\alias{cenken-methods}
\alias{cenken,numeric,logical,numeric,logical-method}
\alias{cenken,numeric,logical,numeric,missing-method}
\alias{cenken,formula,missing,missing,missing-method}
\title{Methods for function cenken in Package NADA}
\description{
    See \code{\link{cenken}} for all the details.
}
\keyword{methods}
