\name{coef-methods}
\docType{methods}
\alias{coef-methods}
\alias{coefficents-methods}
\alias{coef,cenreg-method}
\title{Methods for function coef in package NADA}
\description{
  Methods for extracting coefficients from 
  MLE regression models in package \pkg{NADA} 
}
\usage{
%\S4method{coef}{ros}(object, ...)

\S4method{coef}{cenreg}(object, ...)

} % End usage

\arguments{

\item{object}{
    An output object from a NADA function such as \code{\link{cenreg}}.
}
\item{...}{Additional parameters to subclasses -- currently none}

} % End arguments
\seealso{\code{\link{cenreg}}}

\keyword{methods}
