\name{cor-methods}
\docType{methods}
\alias{cor}
\alias{cor-methods}
\alias{cor,ANY-method}
\alias{cor,cenreg-method}
\title{Methods for function cor in Package NADA}
\description{
  Methods for function \code{cor} in package \pkg{NADA} 
}
\section{Methods}{
\describe{
\item{x = "cenreg"}{Extracts the r-likelihood correlation coefficient from
a \code{cenreg} object.}
}
} % descibe
\keyword{methods}
