\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,ANY-method}
\alias{print,Cen-method}
\alias{print,cenfit-method}
\alias{print,cenmle-method}
\alias{print,ros-method}
\alias{print,NADAlist-method}
\title{Methods for function print in Package NADA}
\description{
  Methods for function \code{print} in package \pkg{NADA} 
}
\section{Methods}{
\describe{

\item{x = "ANY"}{Default print method}

\item{x = "Cen"}{Displays a \code{Cen} object.}

\item{x = "cenfit"}{Displays a \code{cenfit} object.}

\item{x = "cenmle"}{Displays a \code{cenmle} object.}

\item{x = "ros"}{Displays a \code{ros} object.}

%\item{x = "NADAlist"}{Displays a \code{NADAlist} object.  This is an
%internal method and should rarely be used from the command line.}
}}
\keyword{methods}
