\name{residuals-methods}
\docType{methods}
\alias{residuals-methods}
\alias{residuals,cenmle-method}
\title{Methods for function residuals in package NADA}
\description{
  Methods for extracting residuals from 
  MLE regression models in package \pkg{NADA} 
}
\usage{
%\S4method{residuals}{ros}(object, ...)

\S4method{residuals}{cenmle}(object, ...)

}
\arguments{

\item{object}{
    An output object from a NADA function such as \code{\link{cenmle}}.
}
\item{...}{Additional parameters to subclasses -- currently none}

} % End arguments
\seealso{\code{\link{cenmle}}}

\keyword{methods}
