\name{ShePyrene}
\alias{ShePyrene}
\docType{data}
\title{Pyrene concentrations in water from Puget Sound, WA USA}
\usage{data(ShePyrene)}
\source{
She, N., 1997, Analyzing censored water quality data using a nonparametric
approach:  Journal of the American Water Resources Association 33,
615–624..
}
\description{
Pyrene concentrations (mg/L) at 20 monitoring stations in Puget Sound, WA, USA

For characterizing priority pollutant concentrations in sediments of
Puget Sound by computing summary statistics.  Contains eight detection
limits with 11 nondetects out of 56 measurements.
}
\references{
She, N., 1997, Analyzing censored water quality data using a nonparametric
approach:  Journal of the American Water Resources Association 33,
615–624..

Helsel, Dennis R. (2005).  Nondectects and Data Analysis; Statistics
for censored environmental data.  John Wiley and Sons, USA, NJ.
}
\keyword{datasets}


