\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,cenfit-method}
\alias{summary,ros-method}
\title{  Methods for function summary in Package `NADA' }
\description{
  Methods for summarizing objects in package \pkg{NADA} 
}
\usage{
\S4method{summary}{ros}(object, ...)

\S4method{summary}{cenfit}(object, ...)

} % End usage
\arguments{

\item{object}{
    An output object from a NADA fuction such as \code{\link{ros}}.
}
\item{...}{
        Additional arguments passed to the generic method.
}
} % End arguments


\keyword{methods}
